Revert "Merge pull request #19 from fgreinacher/no-activeds"
This reverts commit dbf02e8dc154883c43d66e3d4488e021e6a3eeba, reversing
changes made to 6225ba429a22e08583c20349ec77881a5fc40ca1.
